Jukebox Audio Fixes
Consolidated music from Internal, External, and Mod resources into one music player function instead of different music player functions.

Renamed playTempBgm to playTrack to be less confusing with vanilla

Repathed predefined tracks so they can work with the new player

Added patches for pausing the Jukebox music when a boss fight starts to prevent overlap

Added CampfireExitPatch to prevent background music from turning back on if Jukebox is playing

Added an extra check in the logic to stop the music from playing when you enter rooms with the Jukebox relic

Fixed a minor oversite on a check to stop music when you leave the run.