QoL Updates and Fixes
Anniv7Mod was updated to handle the FadingOut logic and updated Pause logic as JukeboxScreen is culled when not open so it needs to be global

CafeEntryExitPatch was switched from instantly stopping the cafe music to fading out.

JukeboxScreen had some of the update logic moved to Anniv7 for global resources.
Added a Pause function and button
Added a Play function and button

Added text to all the buttons when you hover over them so you can see what they do

Moved skip button down and to the right to be in line with new Pause and Play buttons

Moved the queue buttons slightly left due to new open space from moving Skip button

Added audio checks for the Master_Volume as well as they were ignoring it previously, this will prevent random jarring moments of music changing.

Added FadeOut as a function that can be called to fade the music.

Added new PauseButtonTexture and PlayButtonTexture

Dev Commands Cafe was updated to have the ability to open the jukebox from anywhere and to stop the jukebox which restores the BGM

Added CafeJukebox to handle the dev commands

Updated the relic art to match the new art of the cafe

Added new lines in the UI strings