Major Visual Minor Bug Fixes
Anniv7Mod now has a check to reset the isCoinSlotClicked if a player saves and quits or starts a new run after buying the jukebox.

Added a new visual for Origin/Artist and slightly changed how the files are checked to make sure it accounts for _ and - as well so it can properly check External, Vanilla, or predefined tracks

Adjusted settings on multiple renderings to no longer scale awkwardly between resolutions and now they all scale at the same rates (Primarily the Window was the problem and the queued track numbers)

Made song window slightly larger to help fit longer names

Made the Song buttons slightly bigger to help fit longer names

Adjusted Pause button to light up when active, added checks for hitting skip, queue, and hitting the pause button again to clear the pause state

Made the queue button, confirm button, clear queue, and clear queue off buttons highlight when you hover over them

Added Like A Rolling Stone by Eeriemist as a fun community contribution to the mod.

Fixed the highlighting when a song is playing even if you didn't click on it so when queued songs are played, it will highlight them but only when that page is being rendered.

File changes:
Updated the visuals on the ButtonOn/Off and added ButtonOnGlow/ButtonOffGlow

Added a PauseButtonGlow

Added a new UI string for the Origin/artist field